fix(client): correct fees and reject disconnected requests

+++ xrpl/queries/clio/types/server.go
@@ -39,7 +39,7 @@ type Subscriptions struct {
 // It includes sequence number, hash, fee metrics, and ledger age.
 type LedgerInfo struct {
 	Age            uint          `json:"age"`
-	BaseFeeXRP     float32       `json:"base_fee_xrp"`
+	BaseFeeXRP     float64       `json:"base_fee_xrp"`
 	Hash           types.Hash256 `json:"hash"`
 	ReserveBaseXRP float32       `json:"reserve_base_xrp"`
 	ReserveIncXRP  float32       `json:"reserve_inc_xrp"`
